First working version able to parse input XLSX with Uniprot IDs
Parsing individual IDs specified on command line not tested for
a while.

Stable sort() of the sets and dicts could be introduced to facilitate
testing.

Export to XLSX will be re-enabled, now CSV output is hardcoded.

The most pressing is whether a new column for Uniprot Alias ID should
be introduced, probably yes.

Few more items are exported into CVS/XLSX as python lists which is
not nice.

Could save the sequences into a FASTA file too.