Move info about substrate/intermediate/product/cofactor into dedicated columns
Although those may seem to be ChEBI-related information they are actually not
should be bound to the UniProt-related information about the protein itself.
Also we parse this from the input XLS file so we have to keep the information
associated to a protein+substrate+product combination instead of the respective
chemical substance.

Currently we write multiple Product ChEBI IDs instead of just one into some columns
and a proper condition for the case needs to be crafted. Some attempts at relevant
places are commented out as they still do not work as intended.