Output a single ChEBI substrate/product/etc. in CSV/XLS line per UniProt enzyme
The ChEBI ID column appears twice, the first occurrence can be later dropped, it
is to check whether join between the UniProt vs. ChEBI dicts works properly.

Some columns are totally empty, that's because it seems that the
substrate/intermediate/product/cofactor -related information must be associated
with UniProt-derived entry and able to hold data parsed from previously output CSV/XLS
files. In adeal world maybe when parsing the input one could overwrite the information
parsed from ChEBI with our annotations but one could encounter collisions. Hence,
it seems easier and safer to keep the parsed data along the enzyme properties.
In turn, some of the columns in ChEBI-related dict are now empty.

The substrate, intermediate and cofactor code is maybe not fully enabled or is not
propagated internally, needs more work.