Working version again.
Updated the list of non-terpenes and acyclic terpenes to be ignored
and not to be recognized as products.

The intermediates are not listed as substrates in a whitelist so probably
they end up as a product along with the actual product (this will need
a fix, maybe parsing the reaction string).

A few places are annotated with BUG: or TODO: hints.

The code can process old version of table from Adela from TPS-database_20210517.xlsx
and also from TPS-database_20210623.xlsx.